# gateway/docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API gateway for the BEPetCare system, providing access
    to user, order, product, payment, appointment, and pet record services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BEPetCare Gateway API
  version: "1.0"
paths:
  /appointments:
    post:
      consumes:
      - application/json
      description: Creates a new appointment with details like customer ID, address,
        scheduled time, services, note, and branch ID
      parameters:
      - description: Appointment details
        in: body
        name: request
        required: true
        schema:
          properties:
            branch_id:
              type: integer
            customer_address:
              type: string
            customer_id:
              type: integer
            note:
              type: string
            scheduled_time:
              type: string
            services:
              allOf:
              - type: array
              - properties:
                  quantity:
                    type: integer
                  service_id:
                    type: integer
                type: object
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Appointment created successfully
          schema:
            properties:
              appointment_id:
                type: integer
              status:
                type: string
            type: object
        "400":
          description: Invalid request or invalid scheduled_time format (must be RFC3339)
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new appointment
      tags:
      - Appointments
  /appointments/{appointment_id}:
    get:
      description: Retrieves detailed information for a specific appointment ID, including
        associated order if available
      parameters:
      - description: Appointment ID
        in: path
        name: appointment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Appointment details
          schema:
            properties:
              appointment:
                properties:
                  branch_id:
                    type: integer
                  customer_address:
                    type: string
                  customer_id:
                    type: integer
                  employee_id:
                    type: integer
                  id:
                    type: integer
                  note:
                    type: string
                  scheduled_time:
                    type: string
                  status:
                    type: string
                type: object
              details:
                allOf:
                - type: array
                - properties:
                    quantity:
                      type: integer
                    service_id:
                      type: integer
                  type: object
              order:
                type: object
            type: object
        "400":
          description: Appointment ID is required or invalid format
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Appointment not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get appointment details
      tags:
      - Appointments
  /appointments/customer/{customer_id}:
    get:
      description: Retrieves a list of appointment records for a specific customer
        ID
      parameters:
      - description: Customer ID
        in: path
        name: customer_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of appointments
          schema:
            items:
              properties:
                branch_id:
                  type: integer
                customer_address:
                  type: string
                customer_id:
                  type: integer
                employee_id:
                  type: integer
                id:
                  type: integer
                note:
                  type: string
                scheduled_time:
                  type: string
                status:
                  type: string
              type: object
            type: array
        "400":
          description: Customer ID is required or invalid format
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get appointments by customer
      tags:
      - Appointments
  /appointments/employee/{employee_id}:
    get:
      description: Retrieves a list of appointment records for a specific employee
        ID
      parameters:
      - description: Employee ID
        in: path
        name: employee_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of appointments
          schema:
            items:
              properties:
                branch_id:
                  type: integer
                customer_address:
                  type: string
                customer_id:
                  type: integer
                employee_id:
                  type: integer
                id:
                  type: integer
                note:
                  type: string
                scheduled_time:
                  type: string
                status:
                  type: string
              type: object
            type: array
        "400":
          description: Employee ID is required or invalid format
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get appointments by employee
      tags:
      - Appointments
  /appointments/update-status:
    put:
      consumes:
      - application/json
      description: Updates the status of an appointment for a specific appointment
        ID
      parameters:
      - description: Appointment status update details
        in: body
        name: request
        required: true
        schema:
          properties:
            appointment_id:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Appointment status updated successfully
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Invalid request, invalid appointment_id format, or invalid
            status
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update appointment status
      tags:
      - Appointments
...
  /services:
    get:
      description: Retrieves a list of all service records
      produces:
      - application/json
      responses:
        "200":
          description: List of services
          schema:
            items:
              properties:
                description:
                  type: string
                id:
                  type: integer
                imgurl:
                  type: string
                name:
                  type: string
                price:
                  type: number
              type: object
            type: array
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: List all services
      tags:
      - Services
    post:
      consumes:
      - application/json
      description: Creates a new service with details like name, description, and
        price
      parameters:
      - description: Service details
        in: body
        name: request
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
            price:
              type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Service created successfully
          schema:
            properties:
              service_id:
                type: integer
              status:
                type: string
            type: object
        "400":
          description: Invalid request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create a new service
      tags:
      - Services
    put:
      consumes:
      - application/json
      description: Updates a service with details like service ID, name, description,
        and price
      parameters:
      - description: Updated service details
        in: body
        name: request
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
            price:
              type: number
            service_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Service updated successfully
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Invalid request or invalid service_id format
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update a service
      tags:
      - Services
  /services/{service_id}:
    delete:
      description: Deletes a service by its unique ID
      parameters:
      - description: Service ID
        in: path
        name: service_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Service deleted successfully
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Service ID is required or invalid format
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Delete a service
      tags:
      - Services
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"